+++ code/packages/create-frontend-app/template/+nextjs+azure-web-pubsub/components/actions/color-token-tool/color-token-tool.ts
@@ -8,6 +8,7 @@ import {
 import { COLOR_TOKEN_TOOL_STATE } from './constants';
 import { WeaveAction, WeaveNodesSelectionPlugin } from '@inditextech/weave-sdk';
 import Konva from 'konva';
+import { ColorTokenNode } from '@/components/nodes/color-token/color-token';
 
 export class ColorTokenToolAction extends WeaveAction {
   protected initialized: boolean = false;
@@ -101,7 +102,8 @@ export class ColorTokenToolAction extends WeaveAction {
 
     this.colorTokenId = uuidv4();
 
-    const nodeHandler = this.instance.getNodeHandler('color-token');
+    const nodeHandler =
+      this.instance.getNodeHandler<ColorTokenNode>('color-token');
 
     const node = nodeHandler.create(this.colorTokenId, {
       ...this.props,