+++ code/packages/sdk/src/actions/frame-tool/types.ts
@@ -12,7 +12,15 @@ export type WeaveFrameToolActionState =
 export type WeaveFrameToolCallbacks = WeaveActionCallbacks;
 
 export type WeaveFrameToolActionTriggerParams = {
+  title?: string;
   fontFamily?: string;
+  titleHeight?: number;
+  borderWidth?: number;
+  borderColor?: string;
+  frameWidth?: number;
+  frameHeight?: number;
+  frameOrientation?: string;
+  frameType?: string;
 };
 
 export type WeaveFrameToolProps = {