+++ code/packages/sdk/src/nodes/frame/frame.ts
@@ -10,50 +10,73 @@ import {
   type WeaveStateElement,
   WEAVE_NODE_CUSTOM_EVENTS,
 } from '@inditextech/weave-types';
-import { WEAVE_FRAME_NODE_TYPE } from './constants';
+import {
+  WEAVE_FRAME_NODE_DEFAULT_PROPS,
+  WEAVE_FRAME_NODE_TYPE,
+} from './constants';
+import type { WeaveFrameAttributes } from './types';
 
 export class WeaveFrameNode extends WeaveNode {
   protected nodeType: string = WEAVE_FRAME_NODE_TYPE;
 
-  onRender(props: WeaveElementAttributes): WeaveElementInstance {
-    const { id } = props;
+  create(key: string, props: Partial<WeaveFrameAttributes>): WeaveStateElement {
+    return {
+      key,
+      type: this.nodeType,
+      props: {
+        ...props,
+        id: key,
+        nodeType: this.nodeType,
+        ...WEAVE_FRAME_NODE_DEFAULT_PROPS,
+        ...(props.fontFamily && { title: props.fontFamily }),
+        ...(props.title && { title: props.title }),
+        ...(props.titleHeight && { titleHeight: props.titleHeight }),
+        ...(props.borderColor && { borderColor: props.borderColor }),
+        ...(props.borderWidth && { borderWidth: props.borderWidth }),
+        ...(props.frameWidth && { frameWidth: props.frameWidth }),
+        ...(props.frameHeight && { frameHeight: props.frameHeight }),
+        ...(props.frameType && { frameType: props.frameType }),
+        ...(props.frameOrientation && {
+          frameOrientation: props.frameOrientation,
+        }),
+        children: [],
+      },
+    };
+  }
+
+  onRender(props: WeaveFrameAttributes): WeaveElementInstance {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const { id, zIndex, ...restProps } = props;
 
     const frameParams = {
-      ...props,
+      ...restProps,
     };
-    delete frameParams.fontFamily;
-    delete frameParams.zIndex;
-
-    const frameWidth = 1403;
-    const frameHeight = 992;
-    const titleHeight = 30;
-    const strokeWidth = 2;
 
     const frame = new Konva.Group({
       ...frameParams,
+      id,
       containerId: `${id}-group-internal`,
       containerOffsetX: 0,
-      containerOffsetY: titleHeight + strokeWidth,
-      isTarget: false,
-      width: frameWidth + strokeWidth * 2,
-      height: frameHeight + titleHeight + strokeWidth * 2,
+      containerOffsetY: props.titleHeight + props.borderWidth,
+      width: props.frameWidth + props.borderWidth * 2,
+      height: props.frameHeight + props.titleHeight + props.borderWidth * 2,
       fill: '#ffffffff',
       clipX: 0,
       clipY: 0,
-      clipWidth: frameWidth + strokeWidth * 2,
-      clipHeight: frameHeight + titleHeight + strokeWidth * 2,
+      clipWidth: props.frameWidth + props.borderWidth * 2,
+      clipHeight: props.frameHeight + props.titleHeight + props.borderWidth * 2,
       name: 'node',
     });
 
     const background = new Konva.Rect({
       id: `${id}-bg`,
       nodeId: id,
-      x: strokeWidth,
-      y: titleHeight + strokeWidth,
-      width: frameWidth,
-      stroke: '#000000ff',
-      strokeWidth: 2,
-      height: frameHeight,
+      x: props.borderWidth,
+      y: props.titleHeight + props.borderWidth,
+      width: props.frameWidth,
+      stroke: props.borderColor,
+      strokeWidth: props.borderWidth,
+      height: props.frameHeight,
       fill: '#ffffffff',
       draggable: false,
     });
@@ -64,12 +87,12 @@ export class WeaveFrameNode extends WeaveNode {
       id: `${id}-title`,
       x: 0,
       y: 0,
-      width: frameWidth,
-      height: titleHeight - 10,
+      width: props.frameWidth,
+      height: props.titleHeight - 10,
       fontSize: 20,
       fontFamily: props.fontFamily,
       align: 'left',
-      text: frameParams.title,
+      text: props.title,
       stroke: '#000000ff',
       strokeWidth: 1,
       listening: false,
@@ -81,17 +104,17 @@ export class WeaveFrameNode extends WeaveNode {
     const frameInternal = new Konva.Group({
       id: `${id}-group-internal`,
       nodeId: id,
-      x: strokeWidth,
-      y: titleHeight + strokeWidth,
-      width: frameWidth,
-      height: frameHeight,
+      x: props.borderWidth,
+      y: props.titleHeight + props.borderWidth,
+      width: props.frameWidth,
+      height: props.frameHeight,
       draggable: false,
       stroke: 'transparent',
-      strokeWidth,
+      borderWidth: props.borderWidth,
       clipX: 0,
       clipY: 0,
-      clipWidth: frameWidth,
-      clipHeight: frameHeight,
+      clipWidth: props.frameWidth,
+      clipHeight: props.frameHeight,
     });
 
     frame.add(frameInternal);
@@ -101,7 +124,7 @@ export class WeaveFrameNode extends WeaveNode {
     frame.on(WEAVE_NODE_CUSTOM_EVENTS.onTargetLeave, () => {
       background.setAttrs({
         stroke: '#000000ff',
-        strokeWidth: 2,
+        strokeWidth: props.borderWidth,
         fill: '#ffffffff',
       });
     });
@@ -109,7 +132,7 @@ export class WeaveFrameNode extends WeaveNode {
     frame.on(WEAVE_NODE_CUSTOM_EVENTS.onTargetEnter, () => {
       background.setAttrs({
         stroke: '#ff6863ff',
-        strokeWidth: 2,
+        strokeWidth: props.borderWidth,
         fill: '#ecececff',
       });
     });