+++ code/CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- [#140](https://github.com/InditexTech/weavejs/issues/140) API to customize styling of selection elements
+- [#138](https://github.com/InditexTech/weavejs/issues/138) API to customize styling of frame element
+- [#137](https://github.com/InditexTech/weavejs/issues/137) API to customize styling of selection elements
 
 ## [0.4.0] - 2025-05-05