+++ code/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [#139](https://github.com/InditexTech/weavejs/issues/139) API to obtain a Tree representation of the nodes hierarchy
 - [#138](https://github.com/InditexTech/weavejs/issues/138) API to customize styling of frame element
 - [#137](https://github.com/InditexTech/weavejs/issues/137) API to customize styling of selection elements